requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.5.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-postgres>=0.0.12
pgvector>=0.3.2
psycopg[binary,pool]>=3.2.3
sqlalchemy>=2.0.35
alembic>=1.14.0
python-dotenv>=1.0.1
httpx>=0.27.2
ruff>=0.7.0
PyJWT>=2.12.1
bcrypt<4.0.0
passlib[bcrypt]>=1.7.4
email-validator>=2.3.0